Add user authentication and page access restrictions

+++ base/views.py
@@ -1,15 +1,60 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
 
-# using function based views 
+# using function based views
+
+def login_view(request):
+    
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+            
+        user = authenticate(request, username=username, password=password)
+        
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Credentials do not match')
+            
+    context = {}
+    return render(request, 'base/user_auth.html', context)
+
+
+def logout_user(request):
+    logout(request)
+    return redirect('home')
+
 
 def home(request):
-    rooms = Room.objects.all()
-    topics = Topic.objects.all()
+    # assignment using inline if
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
 
-    context = {'rooms': rooms, 'topics': topics}
+    # case sensitivity for icontains(makes it case insensitive with i included) - for matching url queries
+    # django using & and | instead of 'and' and 'or'
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) |
+        Q(name__icontains=q)        |
+        Q(description__icontains=q) 
+        )
+    topics = Topic.objects.all()
+    room_count = rooms.count()
+
+    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', context)
+
 
 def room_view(request, pk):
     room = Room.objects.get(id=pk)
@@ -18,6 +63,9 @@ def room_view(request, pk):
 
 
 # creating room form input
+# using login_required decorator to limit access to create room functionality 
+
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     
@@ -33,10 +81,14 @@ def create_room(request):
 
 # updating room details
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
     context = {'form': form}
+    
+    if request.user != room.host:
+        return HttpResponse('You are not authorized to manipulate this page!')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -49,8 +101,12 @@ def update_room(request, pk):
 
 # deleting a room
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not authorized to manipulate this page!')
     
     if request.method == 'POST':
         room.delete()
